% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signT.R
\name{Alt}
\alias{Alt}
\title{Alternating optimization of the weighted classification loss}
\usage{
Alt(Ybar,W,r,type = c("logistic","hinge"),start = "random")
}
\arguments{
\item{Ybar}{A given data tensor (possibly noisy and incomplete).}

\item{W}{A weight tensor of the weighted classification loss.}

\item{r}{A rank to be fitted.}

\item{type}{A large margin loss to be used. Logistic or hinge loss is available.}

\item{start}{Choice of initialization method. Use random initialization if \code{start} = "random", the initialization based on low rank approximation if \code{start} = "linear".}
}
\value{
The returned value is a list with components.

\code{binary_obj} \verb{ }A series of binary loss values at each iteration.

\code{obj} \verb{        }A series of the weighted classification loss values at each iteration.

\code{iter} \verb{       }The number of iterations.

\code{error} \verb{      }A series of errors at each iteration.

\code{fitted} \verb{     }A tensor that optimizes the weighted classification loss.
}
\description{
Optimize the weighted classification loss given a weight tensor, an observed data tensor, and a large margin loss.
}
\examples{
library(rTensor)
indices = c(2,3,4)
noise = rand_tensor(indices)@data
Theta = array(runif(prod(indices),min=-3,max = 3),indices)

# The signal plus noise model
Y = Theta + noise

# Optimize the weighted classification for a given sign tensor sign(Y) and a weight tensor abs(Y)
result = Alt(sign(Y),abs(Y),r = 2,type = "logistic",start = "random")
signTheta = sign(result$fitted)

}
\references{
Lee, C., & Wang, M. (2021). Beyond the Signs: Nonparametric Tensor Completion via Sign Series. \emph{arXiv preprint arXiv:2102.00384}.
}
