% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signT.R
\name{fit_continuous}
\alias{fit_continuous}
\title{Estimating a signal tensor from a noisy and incomplete data tensor based on CP low rank tensor method.}
\usage{
fit_continuous(data,r)
}
\arguments{
\item{data}{A given data tensor (possibly noisy and incomplete).}

\item{r}{A rank of the signal tensor.}
}
\value{
The returned value is a list with components.

\code{est}\verb{    } An estimated signal tensor based on CP low rank tensor method.

\code{U}\verb{      } A list of factor matrices.

\code{lambda}\verb{ } A vector of tensor singular values.
}
\description{
Estimate a signal tensor from a noisy and incomplete data tensor using CP low rank tensor method.
}
\examples{
library(rTensor)
indices = c(2,3,4)
noise = rand_tensor(indices)@data
Theta = array(runif(prod(indices),min=-3,max = 3),indices)

# The signal plus noise model
Y = Theta + noise

# Estimate Theta from CP low rank tensor method
hatTheta = fit_continuous(Y,3)
print(hatTheta$est)

}
